moving listen_loiter back to functions.py. Added SleepHelper as a parameter to the Functions constructor
